Allow selecting between LuaJIT and Lua5.4
With this version it introduces a `config.json` with a `luaRuntime` field. By default this is set to LUAJIT which means scripts will be ran under the LUAJIT runtime. Another supported runtime is Lua 5.4 which can be used by setting that field to LUA54.

LuaJIT does require `libgcc` to be present, otherwise an UnsatisfiedLinkError will be thrown however with this version if that occurs it will try and fallback to Lua 5.4 which doesn't seem to require it.